+++ src/biz/bokhorst/xprivacy/ActivityMain.java
@@ -24,7 +24,6 @@ import android.content.res.TypedArray;
 import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.graphics.Typeface;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Build;
@@ -84,7 +83,8 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	private static final int ACTIVITY_EXPORT = 1;
 	private static final int ACTIVITY_IMPORT = 2;
 	private static final int ACTIVITY_IMPORT_SELECT = 3;
-	private static final int ACTIVITY_FETCH = 4;
+	private static final int ACTIVITY_SUBMIT = 4;
+	private static final int ACTIVITY_FETCH = 5;
 
 	private static final int LICENSED = 0x0100;
 	private static final int NOT_LICENSED = 0x0231;
@@ -470,12 +470,25 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 					String fileName = dataIntent.getData().getPath();
 					fileName = fileName.replace("/document/primary:", Environment.getExternalStorageDirectory()
 							.getAbsolutePath() + File.separatorChar);
-					Intent intent = new Intent(ActivityShare.ACTION_IMPORT);
-					intent.putExtra(ActivityShare.cFileName, fileName);
-					startActivityForResult(intent, ACTIVITY_IMPORT);
+					startImport(fileName);
 				} catch (Throwable ex) {
 					Util.bug(null, ex);
 				}
+
+		} else if (requestCode == ACTIVITY_SUBMIT) {
+			// Fetch
+			sharingDone();
+			if (mAppAdapter != null)
+				mAppAdapter.notifyDataSetChanged(); // Update state
+
+			String errorMessage = null;
+			if (dataIntent != null && dataIntent.hasExtra(ActivityShare.cErrorMessage))
+				errorMessage = dataIntent.getStringExtra(ActivityShare.cErrorMessage);
+
+			String text = String.format("%s: %s", getString(R.string.menu_submit),
+					errorMessage == null ? getString(R.string.msg_done) : errorMessage);
+			Toast toast = Toast.makeText(this, text, Toast.LENGTH_LONG);
+			toast.show();
 
 		} else if (requestCode == ACTIVITY_FETCH) {
 			// Fetch
@@ -492,6 +505,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			Toast toast = Toast.makeText(this, text, Toast.LENGTH_LONG);
 			toast.show();
 		}
+
 	}
 
 	@Override
@@ -509,6 +523,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 		menu.findItem(R.id.menu_all).setEnabled(!mBatchOpRunning);
 		menu.findItem(R.id.menu_export).setEnabled(pro && mounted && !mBatchOpRunning);
 		menu.findItem(R.id.menu_import).setEnabled(pro && mounted && !mBatchOpRunning);
+		menu.findItem(R.id.menu_submit).setEnabled(!mBatchOpRunning);
 		menu.findItem(R.id.menu_fetch).setEnabled(!mBatchOpRunning);
 		menu.findItem(R.id.menu_pro).setVisible(!pro);
 
@@ -518,9 +533,31 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	@Override
 	public boolean onOptionsItemSelected(MenuItem item) {
 		try {
+			// Show a dialog if the option needs a selection to work on
+			int id = item.getItemId();
+			if ((id == R.id.menu_all || id == R.id.menu_import || id == R.id.menu_fetch || id == R.id.menu_submit)
+					&& mAppAdapter != null && mAppAdapter.getSelected().size() <= 0) {
+				AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+				alertDialogBuilder.setTitle(R.string.app_name);
+				alertDialogBuilder.setMessage(getString(R.string.msg_select));
+				alertDialogBuilder.setIcon(Util.getThemed(this, R.attr.icon_launcher));
+				alertDialogBuilder.setPositiveButton(getString(android.R.string.ok),
+						new DialogInterface.OnClickListener() {
+							@Override
+							public void onClick(DialogInterface dialog, int which) {
+							}
+						});
+				AlertDialog alertDialog = alertDialogBuilder.create();
+				alertDialog.show();
+				return true;
+			}
+
 			switch (item.getItemId()) {
 			case R.id.menu_help:
 				optionHelp();
+				return true;
+			case R.id.menu_select_all:
+				optionSelectAll();
 				return true;
 			case R.id.menu_tutorial:
 				optionTutorial();
@@ -545,6 +582,9 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 				return true;
 			case R.id.menu_import:
 				optionImport();
+				return true;
+			case R.id.menu_submit:
+				optionSubmit();
 				return true;
 			case R.id.menu_fetch:
 				optionFetch();
@@ -660,43 +700,46 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	// Options
 
 	private void optionAll() {
-		// Check if some restricted
-		boolean some = false;
-		for (int pos = 0; pos < mAppAdapter.getCount(); pos++) {
-			ApplicationInfoEx xAppInfo = mAppAdapter.getItem(pos);
-			for (boolean restricted : PrivacyManager.getRestricted(ActivityMain.this, xAppInfo.getUid(),
-					mAppAdapter.getRestrictionName()))
-				if (restricted) {
-					some = true;
+		if (mAppAdapter != null) {
+			// Check if some restricted
+			boolean some = false;
+			for (ApplicationInfoEx xAppInfo : mAppAdapter.getSelected()) {
+				for (boolean restricted : PrivacyManager.getRestricted(ActivityMain.this, xAppInfo.getUid(),
+						mAppAdapter.getRestrictionName()))
+					if (restricted) {
+						some = true;
+						break;
+					}
+				if (some)
 					break;
-				}
-			if (some)
-				break;
-		}
-		final boolean someRestricted = some;
+			}
+			final boolean someRestricted = some;
 
-		// Are you sure?
-		AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
-		alertDialogBuilder.setTitle(getString(someRestricted ? R.string.menu_clear_all : R.string.menu_restrict_all));
-		alertDialogBuilder.setMessage(getString(R.string.msg_sure));
-		alertDialogBuilder.setIcon(Util.getThemed(this, R.attr.icon_launcher));
-		alertDialogBuilder.setPositiveButton(getString(android.R.string.ok), new DialogInterface.OnClickListener() {
-			@Override
-			public void onClick(DialogInterface dialog, int which) {
-				if (mAppAdapter != null) {
-					mBatchOpRunning = true;
-					invalidateOptionsMenu();
-					new ToggleTask().executeOnExecutor(mExecutor, someRestricted);
+			// Are you sure?
+			AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+			alertDialogBuilder
+					.setTitle(getString(someRestricted ? R.string.menu_clear_all : R.string.menu_restrict_all));
+			alertDialogBuilder.setMessage(getString(R.string.msg_sure));
+			alertDialogBuilder.setIcon(Util.getThemed(this, R.attr.icon_launcher));
+			alertDialogBuilder.setPositiveButton(getString(android.R.string.ok), new DialogInterface.OnClickListener() {
+				@Override
+				public void onClick(DialogInterface dialog, int which) {
+					if (mAppAdapter != null) {
+						mBatchOpRunning = true;
+						invalidateOptionsMenu();
+						new ToggleTask().executeOnExecutor(mExecutor, someRestricted);
+					}
 				}
-			}
-		});
-		alertDialogBuilder.setNegativeButton(getString(android.R.string.cancel), new DialogInterface.OnClickListener() {
-			@Override
-			public void onClick(DialogInterface dialog, int which) {
-			}
-		});
-		AlertDialog alertDialog = alertDialogBuilder.create();
-		alertDialog.show();
+			});
+			alertDialogBuilder.setNegativeButton(getString(android.R.string.cancel),
+					new DialogInterface.OnClickListener() {
+						@Override
+						public void onClick(DialogInterface dialog, int which) {
+						}
+					});
+			AlertDialog alertDialog = alertDialogBuilder.create();
+			alertDialog.show();
+		}
 	}
 
 	private void optionUsage() {
@@ -770,10 +813,67 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			chooseFile.setDataAndType(uri, "text/xml");
 			Intent intent = Intent.createChooser(chooseFile, getString(R.string.app_name));
 			startActivityForResult(intent, ACTIVITY_IMPORT_SELECT);
+		} else
+			startImport(ActivityShare.getFileName(false));
+	}
+
+	private void startImport(String fileName) {
+		Intent intent = new Intent(ActivityShare.ACTION_IMPORT);
+		intent.putExtra(ActivityShare.cFileName, fileName);
+		int[] uid;
+		if (mAppAdapter == null)
+			uid = new int[0];
+		else {
+			List<ApplicationInfoEx> listAppInfo = mAppAdapter.getSelected();
+			uid = new int[listAppInfo.size()];
+			for (int pos = 0; pos < listAppInfo.size(); pos++)
+				uid[pos] = listAppInfo.get(pos).getUid();
+			intent.putExtra(ActivityShare.cUidList, uid);
+		}
+		startActivityForResult(intent, ACTIVITY_IMPORT);
+	}
+
+	private void optionSubmit() {
+		if (mAppAdapter.getSelected().size() <= ActivityShare.cSubmitLimit) {
+			AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+			alertDialogBuilder.setTitle(getString(R.string.menu_submit));
+			alertDialogBuilder.setMessage(getString(R.string.msg_sure));
+			alertDialogBuilder.setIcon(Util.getThemed(this, R.attr.icon_launcher));
+			alertDialogBuilder.setPositiveButton(getString(android.R.string.ok), new DialogInterface.OnClickListener() {
+				@Override
+				public void onClick(DialogInterface dialog, int which) {
+					sharingStart();
+					if (mAppAdapter != null) {
+						List<ApplicationInfoEx> listAppInfo = mAppAdapter.getSelected();
+						int[] uid = new int[listAppInfo.size()];
+						for (int pos = 0; pos < listAppInfo.size(); pos++)
+							uid[pos] = listAppInfo.get(pos).getUid();
+						Intent intent = new Intent(ActivityShare.ACTION_SUBMIT);
+						intent.putExtra(ActivityShare.cUidList, uid);
+						startActivityForResult(intent, ACTIVITY_SUBMIT);
+					}
+				}
+			});
+			alertDialogBuilder.setNegativeButton(getString(android.R.string.cancel),
+					new DialogInterface.OnClickListener() {
+						@Override
+						public void onClick(DialogInterface dialog, int which) {
+						}
+					});
+			AlertDialog alertDialog = alertDialogBuilder.create();
+			alertDialog.show();
 		} else {
-			Intent intent = new Intent(ActivityShare.ACTION_IMPORT);
-			intent.putExtra(ActivityShare.cFileName, ActivityShare.getFileName(false));
-			startActivityForResult(intent, ACTIVITY_IMPORT);
+			AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+			alertDialogBuilder.setTitle(getString(R.string.app_name));
+			alertDialogBuilder.setMessage(getString(R.string.msg_limit, ActivityShare.cSubmitLimit));
+			alertDialogBuilder.setIcon(Util.getThemed(this, R.attr.icon_launcher));
+			alertDialogBuilder.setPositiveButton(getString(android.R.string.ok), new DialogInterface.OnClickListener() {
+				@Override
+				public void onClick(DialogInterface dialog, int which) {
+				}
+			});
+			AlertDialog alertDialog = alertDialogBuilder.create();
+			alertDialog.show();
 		}
 	}
 
@@ -791,8 +891,15 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 				@Override
 				public void onClick(DialogInterface dialog, int which) {
 					sharingStart();
-					Intent intent = new Intent(ActivityShare.ACTION_FETCH);
-					startActivityForResult(intent, ACTIVITY_FETCH);
+					if (mAppAdapter != null) {
+						List<ApplicationInfoEx> listAppInfo = mAppAdapter.getSelected();
+						int[] uid = new int[listAppInfo.size()];
+						for (int pos = 0; pos < listAppInfo.size(); pos++)
+							uid[pos] = listAppInfo.get(pos).getUid();
+						Intent intent = new Intent(ActivityShare.ACTION_FETCH);
+						intent.putExtra(ActivityShare.cUidList, uid);
+						startActivityForResult(intent, ACTIVITY_FETCH);
+					}
 				}
 			});
 			alertDialogBuilder.setNegativeButton(getString(android.R.string.cancel),
@@ -869,6 +976,12 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 		imgHelpHalf.setImageBitmap(mCheck[1]);
 		dialog.setCancelable(true);
 		dialog.show();
+	}
+
+	private void optionSelectAll() {
+		// Select all visible apps
+		if (mAppAdapter != null)
+			mAppAdapter.selectAllVisible();
 	}
 
 	private void optionTutorial() {
@@ -1013,19 +1126,22 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 
 			// Apply action
 			boolean restart = false;
-			for (int pos = 0; pos < mAppAdapter.getCount(); pos++) {
-				publishProgress(pos, mAppAdapter.getCount());
-				ApplicationInfoEx xAppInfo = mAppAdapter.getItem(pos);
-				if (mAppAdapter.getRestrictionName() == null && someRestricted)
-					restart = PrivacyManager.deleteRestrictions(ActivityMain.this, xAppInfo.getUid()) || restart;
-				else if (mAppAdapter.getRestrictionName() == null) {
-					for (String restrictionName : PrivacyManager.getRestrictions())
+			if (mAppAdapter != null) {
+				int pos = 0;
+				List<ApplicationInfoEx> listAppInfo = mAppAdapter.getSelected();
+				for (ApplicationInfoEx xAppInfo : listAppInfo) {
+					publishProgress(pos++, listAppInfo.size());
+					if (mAppAdapter.getRestrictionName() == null && someRestricted)
+						restart = PrivacyManager.deleteRestrictions(ActivityMain.this, xAppInfo.getUid()) || restart;
+					else if (mAppAdapter.getRestrictionName() == null) {
+						for (String restrictionName : PrivacyManager.getRestrictions())
+							restart = PrivacyManager.setRestricted(null, ActivityMain.this, xAppInfo.getUid(),
+									restrictionName, null, !someRestricted) || restart;
+					} else
 						restart = PrivacyManager.setRestricted(null, ActivityMain.this, xAppInfo.getUid(),
-								restrictionName, null, !someRestricted) || restart;
-				} else
-					restart = PrivacyManager.setRestricted(null, ActivityMain.this, xAppInfo.getUid(),
-							mAppAdapter.getRestrictionName(), null, !someRestricted)
-							|| restart;
+								mAppAdapter.getRestrictionName(), null, !someRestricted)
+								|| restart;
+				}
 			}
 
 			return restart;
@@ -1066,7 +1182,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	@SuppressLint("DefaultLocale")
 	private class AppListAdapter extends ArrayAdapter<ApplicationInfoEx> {
 		private Context mContext;
-		private List<ApplicationInfoEx> mListApp;
+		private List<ApplicationInfoEx> mListAppAll;
 		private List<ApplicationInfoEx> mListAppSelected = new ArrayList<ApplicationInfoEx>();
 		private String mRestrictionName;
 		private LayoutInflater mInflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
@@ -1077,8 +1193,8 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 				String initialRestrictionName) {
 			super(context, resource, objects);
 			mContext = context;
-			mListApp = new ArrayList<ApplicationInfoEx>();
-			mListApp.addAll(objects);
+			mListAppAll = new ArrayList<ApplicationInfoEx>();
+			mListAppAll.addAll(objects);
 			mRestrictionName = initialRestrictionName;
 
 			TypedArray ta1 = context.getTheme().obtainStyledAttributes(
@@ -1093,6 +1209,41 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 
 		public String getRestrictionName() {
 			return mRestrictionName;
+		}
+
+		public List<ApplicationInfoEx> getSelected() {
+			return mListAppSelected;
+		}
+
+		public void selectAllVisible() {
+			// Look through the visible apps to figure out what to do
+			boolean addThem = false;
+			for (int i = 0; i < this.getCount(); i++) {
+				if (!mListAppSelected.contains(this.getItem(i))) {
+					addThem = true;
+					break;
+				}
+			}
+
+			if (addThem) {
+				// Add the visible apps not already selected
+				for (int i = 0; i < this.getCount(); i++)
+					if (!mListAppSelected.contains(this.getItem(i)))
+						mListAppSelected.add(this.getItem(i));
+			} else {
+				mListAppSelected.clear();
+			}
+
+			this.showStats();
+			this.notifyDataSetChanged();
+		}
+
+		public void showStats() {
+			TextView tvStats = (TextView) findViewById(R.id.tvStats);
+			String stats = String.format("%d/%d", this.getCount(), mListAppAll.size());
+			if (mListAppSelected.size() > 0)
+				stats += String.format(" (%d)", mListAppSelected.size());
+			tvStats.setText(stats);
 		}
 
 		@Override
@@ -1122,9 +1273,9 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 
 				// Match applications
 				int current = 0;
-				int max = AppListAdapter.this.mListApp.size();
+				int max = AppListAdapter.this.mListAppAll.size();
 				List<ApplicationInfoEx> lstApp = new ArrayList<ApplicationInfoEx>();
-				for (ApplicationInfoEx xAppInfo : AppListAdapter.this.mListApp) {
+				for (ApplicationInfoEx xAppInfo : AppListAdapter.this.mListAppAll) {
 					// Check if another filter has been started
 					if (filtersRunning != mFiltersRunning.get())
 						return null;
@@ -1208,7 +1359,6 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 					ProgressBar pbFilter = (ProgressBar) findViewById(R.id.pbFilter);
 					pbFilter.setVisibility(ProgressBar.GONE);
 					tvStats.setVisibility(TextView.VISIBLE);
-					tvStats.setText(results.count + "/" + AppListAdapter.this.mListApp.size());
 
 					Intent progressIntent = new Intent(ActivityShare.cProgressReport);
 					progressIntent.putExtra(ActivityShare.cProgressMessage, getString(R.string.title_restrict));
@@ -1226,6 +1376,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 						addAll((ArrayList<ApplicationInfoEx>) results.values);
 						notifyDataSetChanged();
 					}
+					AppListAdapter.this.showStats();
 				}
 			}
 		}
@@ -1234,6 +1385,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			private View row;
 			private int position;
 			public View vwState;
+			public LinearLayout llAppType;
 			public ImageView imgIcon;
 			public ImageView imgUsed;
 			public ImageView imgGranted;
@@ -1247,6 +1399,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 				row = theRow;
 				position = thePosition;
 				vwState = (View) row.findViewById(R.id.vwState);
+				llAppType = (LinearLayout) row.findViewById(R.id.llAppType);
 				imgIcon = (ImageView) row.findViewById(R.id.imgIcon);
 				imgUsed = (ImageView) row.findViewById(R.id.imgUsed);
 				imgGranted = (ImageView) row.findViewById(R.id.imgGranted);
@@ -1314,6 +1467,13 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			@Override
 			protected void onPostExecute(Object result) {
 				if (holder.position == position && result != null) {
+					// Set background color
+					if (xAppInfo.isSystem())
+						holder.llAppType.setBackgroundColor(getResources().getColor(
+								Util.getThemed(ActivityMain.this, R.attr.color_dangerous)));
+					else
+						holder.llAppType.setBackgroundColor(Color.TRANSPARENT);
+
 					// Display state
 					if (state == STATE_ATTENTION)
 						holder.vwState.setBackgroundColor(getResources().getColor(
@@ -1372,6 +1532,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 							else
 								holder.row.setBackgroundColor(Color.TRANSPARENT);
 
+							AppListAdapter.this.showStats();
 							return true;
 						}
 					});
@@ -1486,14 +1647,6 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 
 			// Get info
 			final ApplicationInfoEx xAppInfo = getItem(holder.position);
-			holder.imgIcon.setVisibility(View.INVISIBLE);
-
-			// Set background color
-			if (xAppInfo.isSystem())
-				holder.rlName.setBackgroundColor(getResources().getColor(
-						Util.getThemed(ActivityMain.this, R.attr.color_dangerous)));
-			else
-				holder.rlName.setBackgroundColor(Color.TRANSPARENT);
 
 			// Handle details click
 			holder.imgIcon.setOnClickListener(new View.OnClickListener() {
@@ -1508,7 +1661,10 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			});
 
 			// Set data
+			holder.row.setBackgroundColor(Color.TRANSPARENT);
 			holder.vwState.setBackgroundColor(Color.TRANSPARENT);
+			holder.llAppType.setBackgroundColor(Color.TRANSPARENT);
+			holder.imgIcon.setVisibility(View.INVISIBLE);
 			holder.tvName.setText(xAppInfo.toString());
 			holder.tvName.setTypeface(null, Typeface.NORMAL);
 			holder.imgUsed.setVisibility(View.INVISIBLE);